Refactored deployments. Separated each non-standard command into its own file. Added the ability to pass in multiple --component flags to support N number of components for create/override.
